onboarding_steps: Rename 'hotspots' files to 'onboarding_steps'.
