+++ zerver/views/onboarding_steps.py
@@ -1,10 +1,10 @@
 from django.http import HttpRequest, HttpResponse
 from django.utils.translation import gettext as _
 
-from zerver.actions.hotspots import do_mark_onboarding_step_as_read
+from zerver.actions.onboarding_steps import do_mark_onboarding_step_as_read
 from zerver.decorator import human_users_only
 from zerver.lib.exceptions import JsonableError
-from zerver.lib.hotspots import ALL_ONBOARDING_STEPS
+from zerver.lib.onboarding_steps import ALL_ONBOARDING_STEPS
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
 from zerver.models import UserProfile
